+++ WHATSNEW.md
@@ -2,5 +2,6 @@
 
 ### Changed
 
-### Removed
+- ems get workgroups now displays group id in table
 
+### Removed